requirements.txt
Flask==2.3.2
Flask-SQLAlchemy==3.0.5
ultralytics
opencv-python==4.8.0.74
Pillow==10.0.0
python-dotenv==1.0.0
PyYAML==6.0.1
requests==2.31.0
numpy>=1.21,<2
psycopg2-binary>=2.9.11
boxmot>=16.0.8
openpyxl>=3.1.5